Created fixture for DelegateClock and TestingRpcService.
