+++ mantis-common/src/testFixtures/java/io/mantisrx/common/util/DelegateClock.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2022 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.mantisrx.common.util;
+
+import java.time.Clock;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.util.concurrent.atomic.AtomicReference;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class DelegateClock extends Clock {
+    private final AtomicReference<Clock> delegate;
+
+    @Override
+    public ZoneId getZone() {
+        return delegate.get().getZone();
+    }
+
+    @Override
+    public Clock withZone(ZoneId zone) {
+        return delegate.get().withZone(zone);
+    }
+
+    @Override
+    public Instant instant() {
+        return delegate.get().instant();
+    }
+}
