+++ mantis-control-plane/mantis-control-plane-core/build.gradle
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+apply plugin: 'java-test-fixtures'
 
 ext {
