+++ mantis-control-plane/mantis-control-plane-core/src/testFixtures/java/io/mantisrx/server/core/TestingRpcService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.mantisrx.server.worker;
+package io.mantisrx.server.core;
 
 import static org.apache.flink.util.Preconditions.checkNotNull;
