+++ mantis-control-plane/mantis-control-plane-server/build.gradle
@@ -48,6 +48,8 @@ dependencies {
     testCompile libraries.junit4
     testCompile libraries.mockitoAll
     testCompile "org.testng:testng:$testngVersion"
+    testImplementation testFixtures(project(":mantis-common"))
+    testImplementation testFixtures(project(":mantis-control-plane:mantis-control-plane-core"))
 }
 
 mainClassName = "io.mantisrx.server.master.MasterMain"
