+++ mantis-control-plane/mantis-control-plane-server/src/main/java/io/mantisrx/master/resourcecluster/ResourceClusterActor.java
@@ -432,8 +432,9 @@ class ResourceClusterActor extends AbstractActor {
         List<TaskExecutorID> taskExecutors;
     }
 
+    @SuppressWarnings("UnusedReturnValue")
     @AllArgsConstructor
-    private static class TaskExecutorState {
+    static class TaskExecutorState {
 
         enum RegistrationState {
             Registered,
